requirements.txt
numpy~=2.2.4
pandas~=2.2.3
python-dotenv~=1.1.0
lseg-data~=2.1.0
DatastreamPy~=2.0.30
openpyxl~=3.2.0b1